requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.1.1
python-dotenv